finish markup for linter tab
